scripts/podcast-search-mcp-server/pyproject.toml
[project]
name = "podcast-search-mcp-server"
version = "1.0.0"
description = "MCP server for searching AWS French Podcast episodes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "boto3>=1.34.0",
    "feedparser>=6.0.0",
    "python-dateutil>=2.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.100.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
